+++ mrz-3-edition/tester.m
@@ -53,10 +53,10 @@ parabolaSolverComplexVersion(9, 1, 1);  % -0.0556 - 0.3287i, -0.0556 + 0.3287i
 % algorithms
 fixedPoint(@sin, 0.5, 10^(-8), 100);  % 0.1627, 100
 newton(@sin, @cos, 1, 10^(-8), 100);  % 0, 5, false
-multipleRootsNewton(@(x) x.^2, @(x) 2 * x, 1, 10^(-16), 100, 1);  % 5.551115123125783e-17, 54
+multipleRootsNewton(@(x) x.^2, @(x) 2 * x, 3, 10^(-16), 100, 1);  % 5.551115123125783e-17, 54
 halley(@(x) -5 * x.^4 + 6, @(x) -20 * x.^3, @(x) -60 * x.^2, 1, 10^(-16), 100);  % 1.046635139392106, 4
-[x, p, c, numberOfIterations] = convergenceEvaluation(10 ^ (-16), 100);  % 1.0e+10, 2.0, 1.34e+5, 73
-% multiplicityEvaluation();
+convergenceEvaluation(10 ^ (-16), 100);  % 1.0e+10, 2.0, 1.34e+5, 73
+multiplicityEvaluation(@(x) x.^2, @(x) 2 * x, 1, 10^(-16), 100);  % 5.551115123125783e-17, 0.5, 2, 54
 % laguerre();
 % bairstow();